ZEN-28901: update graph data despite bad Predictive Threshold
If there was a datapoint in your predictive threshold that is not
on the Graph, it won't update when you change the range of time.
Update the Graph despite the fact that there is a bad Predictive
Threshold on it. There is no harm in having that Threshold on it.